requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
alembic
passlib[bcrypt]
python-jose[cryptography]
python-multipart